Check for .vagrant before deleting it

+++ vagrant_update.sh
@@ -1,7 +1,10 @@
 #!/bin/env bash
 set -Eeuxo pipefail
 read -p "Enter your Windows username: " WINDOWS_USERNAME
-rm -rf .vagrant
+if [ -d .vagrant ]; then
+    echo "Removing .vagrant directory..."
+    rm -rf .vagrant
+fi
 WINDOWS_PATH="/mnt/c/Users/$WINDOWS_USERNAME/Desktop/k8s-project/.vagrant"
 echo "Copying .vagrant directory from Windows to Linux..."
 cp -rv "$WINDOWS_PATH" ./
